feat: Add a pane grid mechanism
This allows adding multiple tools, and organizing the workspace as you
want; The code is still a bit messy, the intention would be to pull some
of these things into sub-components, and abstract away some of the
boilerplate.

It's called tabbed_pane_grid because I also want to add the ability to
have multiple tabs, etc.